#1 쿼리 작성중

+++ src/main/java/com/playjo/project/playjo/domain/Place.java
@@ -42,7 +42,7 @@ public class Place {
     private Long readCount;
 
     @Column(name = "rating_Avg")
-    private int ratingAvg;
+    private Double ratingAvg;
 
     @ManyToOne
     @JoinColumn(name = "category_no")
